fix(Picker): extracting methods to the usePickerItems
